dom/tests/mochitest/bugs/test_bug407839.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=407839
-->
<head>
  <title>Test for Bug 407839</title>
  <meta http-equiv="Content-type" content="text/html;charset=UTF-8" />
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=407839">Mozilla Bug 407839</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<iframe name="child" src="http://TEST1.example.org/tests/dom/tests/mochitest/bugs/iframe_bug407839-1.html"></iframe>
<iframe name="idn" src="http://sub1.ält.example.org:8000/tests/dom/tests/mochitest/bugs/iframe_bug407839-2.html"></iframe>
<pre id="test">
<script class="testbody" type="text/javascript">
/** Test for Bug 407839 **/

SimpleTest.waitForExplicitFinish();

var gNumMessages = 0;

function receiveMessage(evt)
{
  is(evt.data, "child-response", "got wrong response");

  if (++gNumMessages == 2) {
    SimpleTest.finish();
  }
}

window.addEventListener("message", receiveMessage, false);

</script>
</pre>
</body>
</html>
